tests/e2e/requirements.txt
kafka-python>=2.0.2
psycopg2-binary>=2.9.6
requests>=2.28.0
